Move code around to decrease complexity
