+++ packages/apolloschurchapp/src/notifications/index.js
@@ -1,0 +1,1 @@
+export NotificationsManager from './NotificationManager';
